Reuse fallback ahead count for default-branch status
- Reuse the no-upstream ahead count when comparing against the default branch
- Add coverage for feature branches without upstreams
- Move source control provider icons into shared exports
